+++ src/brasil/gov/agenda/tests/test_compromisso.py
@@ -72,24 +72,6 @@ class ContentTypeTestCase(unittest.TestCase):
     def test_exclude_from_nav_behavior(self):
         self.assertFalse(IExcludeFromNavigation.providedBy(self.compromisso))
 
-    def test_subjects_catalog(self):
-        compromisso = self.compromisso
-        compromisso.subjects = (u'Brasil', u'México')
-        compromisso.reindexObject(idxs=['Subject'])
-        ct = self.portal.portal_catalog
-        results = ct.searchResults(portal_type='Compromisso')
-        b = results[0]
-        self.assertIn('Brasil', b.Subject)
-        self.assertIn('México', b.Subject)
-
-    def test_default_subjects(self):
-        from brasil.gov.agenda.content.compromisso import default_subjects
-        agendadiaria = self.agendadiaria
-        agendadiaria.subjects = (u'Brasil', u'México')
-        # default_factory é executado no container
-        self.assertIn(u'Brasil', default_subjects(agendadiaria))
-        self.assertIn(u'México', default_subjects(agendadiaria))
-
     def test_default_start_date(self):
         from brasil.gov.agenda.content.compromisso import default_start_date
         date_fmt = '%Y-%m-%d %H:%M'