+++ src/brasil/gov/agenda/browser/agendadiaria.py
@@ -92,6 +92,7 @@ class AgendaDiariaView(BrowserView, AgendaMixin):
             comp = {}
             comp['autoridade'] = self.autoridade()
             comp['title'] = obj.Title()
+            comp['solicitante'] = obj.solicitante
             comp['description'] = obj.Description()
             comp['start_time'] = obj.start_date.strftime('%Hh%M')
             comp['start_date'] = obj.start_date.strftime('%Y-%m-%d %H:%M')